refactor: Moves works, instances, and resources routes to separate modules
